+++ src/inject.ts
@@ -1,22 +1,39 @@
-import { dependenciesSymbol, inheritancePreserveSymbol } from './symbols';
-import { Dependency, ImplementationType } from './types';
+import { InheritancePreserver } from './InheritancePreserver';
+import { dependenciesSymbol, injectionSymbol } from './symbols';
+import { Constructor, Dependency, ImplementationType, ImplementationTypeWithInjection } from './types';
 
-function inject<TClass extends new(...args: any[]) => any, TDependency extends Dependency>(dependency: TDependency) {
+function inject<TClass extends Constructor, TDependency extends Dependency>(dependency: TDependency) {
   return function (constructor: TClass, propertyKey: string | symbol, parameterIndex: number) {
     const implementation = constructor as ImplementationType<TClass>;
 
     // if class inherited from another one the static property will be contain the same array ref,
     // and we shouldn't change dependencies of base class
-    const lastInjectedConstructor = implementation[inheritancePreserveSymbol];
-    const injectInNewClass = lastInjectedConstructor !== constructor;
+    const modifiedConstructor = InheritancePreserver.getModifiedConstructor(constructor);
+    const injectInNewClass = modifiedConstructor !== constructor;
 
     if (!implementation[dependenciesSymbol] || injectInNewClass) {
       implementation[dependenciesSymbol] = [];
     }
     implementation[dependenciesSymbol]![parameterIndex] = dependency;
 
-    implementation[inheritancePreserveSymbol] = constructor;
+    InheritancePreserver.constructorModified(constructor);
   };
 }
 
-export { inject };
+function setInjectedParams<TInstance>(implementationType: ImplementationType<TInstance>, injectionParams: any[]) {
+  (implementationType as ImplementationTypeWithInjection<TInstance>)[injectionSymbol] = injectionParams;
+  InheritancePreserver.constructorModified(implementationType as Constructor);
+}
+
+function getInjectedParams<TClass extends Constructor>(constructor: TClass): any[] {
+  const implementation = constructor as ImplementationTypeWithInjection<TClass>;
+  const modifiedConstructor = InheritancePreserver.getModifiedConstructor(constructor);
+  const injected = implementation[injectionSymbol];
+  if (!modifiedConstructor || modifiedConstructor !== constructor || !Array.isArray(injected)) {
+    return [];
+  }
+
+  return injected!;
+}
+
+export { inject, setInjectedParams, getInjectedParams };